Ensure that all archive filenames are valid for PULP
Pulp cannot deal with filenames with comma chars in the filename as it is used
for the PULP_MANIFEST delimiter. Changing pulp-server to use `.rsplit()` rather
then `.split()` on all supported deployments is somewhat unlikely at this point.

Also, ensure the filename is restricted to ASCII encoding, as mirror targets
might be on filesystems that do not handle all of UTF-8 correctly.

This affects exactly one firmware on the LVFS.