Always set the Component.verfmt explicitly on all firmware
Rather than getting the correct version format by falling back to the protocol
or vendor default, allow QA users to explicitly set the value on the LVFS.

This allows the user to initially set or change the verfmt post-upload, for
instance changing it from 'pair' to 'bcd'. This also allows us to support
devices that report the firmware in runtime mode in a different format to what
the protocol is expecting.

This is the case for the Thelio device which reports a triplet in runtime mode
from reading a special device file, but flashes over DFU which is bcd only.