Turn the dropped-GUID from an upload flash() to a waivable test
Too many vendors are ignoring the warnings when uploading firmware with a bad
'generic' AppStream ID and then chaos happens when more than one model is
found in testing or stable.

Move this test to an actual LVFS Test, as it can be re-run as required and
waived by the QA user.