Assign a release tag style for specific vendor per category
This would, for instance, allow us to enforce that TBXXXXXX should be used for
all Lenovo ThunderBolt firmware, or that SWDXXXXX should be advised for all
Dell EC firmware.

Provide a catch-all tag style for the vendor so that we can show the correct
name in the component view.

The tag contents are not currently validated, and no other fields are populated
based on the tag -- yet. Futher work to do.